--- src/main/java/com/example/cloudservice/model/Role.java
@@ -1,7 +1,0 @@
-package com.example.cloudservice.model;
-
-public enum Role {
-
-    USER,
-    ADMIN
-}
