+++ src/main/java/com/example/cloudservice/model/authentication/Role.java
@@ -1,0 +1,7 @@
+package com.example.cloudservice.model.authentication;
+
+public enum Role {
+
+    USER,
+    ADMIN
+}
