+++ src/main/java/com/example/cloudservice/model/token/TokenType.java
@@ -1,0 +1,5 @@
+package com.example.cloudservice.model.token;
+
+public enum TokenType {
+    BEARER
+}
