--- src/main/java/com/example/cloudservice/token/TokenType.java
@@ -1,5 +1,0 @@
-package com.example.cloudservice.token;
-
-public enum TokenType {
-    BEARER
-}
